move WorkerDateProvider.ts to common
