+++ src/api/worker/WorkerLocator.ts
@@ -36,7 +36,7 @@ import {EphemeralCacheStorage} from "./rest/EphemeralCacheStorage"
 import {AdminClientDummyEntityRestCache} from "./rest/AdminClientDummyEntityRestCache.js"
 import {SleepDetector} from "./utils/SleepDetector.js"
 import {SchedulerImpl} from "../common/utils/Scheduler.js"
-import {WorkerDateProvider} from "./utils/WorkerDateProvider.js"
+import {WorkerDateProvider} from "../common/utils/WorkerDateProvider.js"
 import {CacheStorageLateInitializer, LateInitializedCacheStorageImpl} from "./rest/CacheStorageProxy"
 import {IServiceExecutor} from "../common/ServiceRequest"
 import {ServiceExecutor} from "./rest/ServiceExecutor"
